modif: created a componant to avoid replication of code and added tint color for apple user
